requirements.txt
Flask==0.10.1
Werkzeug==0.11.3
facebook-sdk==0.4.0
itsdangerous==0.24
oauthlib==1.0.3
requests-oauthlib==0.6.0
tweepy==3.5.0
